Fixing the links in the readme that go to the audio and video answers (#511)
Github automatically puts "is:open" along with the "label:audio-answer" with the current links in the readme, but once you answer the questions you close the issue, so nothing shows up. 

Specifying the query in the url and not including the "is" field shows any open or closed issues with the corresponding label.